requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
blinker==1.9.0
cachetools==6.2.1
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
gitdb==4.0.12
gitpython==3.1.45
google-auth==2.42.1
google-genai==1.47.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
markupsafe==3.0.3
narwhals==2.10.0
numpy==2.3.4
packaging==25.0
pandas==2.3.3
pillow==12.0.0
protobuf==5.29.5
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1-modules==0.4.2
pydantic==2.12.3
pydantic-core==2.41.4
pydeck==0.9.1
python-dateutil==2.9.0.post0
pytz==2025.2
referencing==0.37.0
requests==2.32.5
rpds-py==0.28.0
rsa==4.9.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
streamlit==1.51.0
tenacity==9.1.2
toml==0.10.2
tornado==6.5.2
typing-extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.2
urllib3==2.5.0
watchdog==6.0.0
websockets==15.0.1
